requirements.txt
black==23.7.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
coloredlogs==15.0.1
humanfriendly==10.0
idna==3.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.4
